22 - Redesigned Audio
Made a base audio scriptable object and a simple auido event class that allows for quick and easy making of new SFX. I also made the Audio now be handled by an Audio manager script instead of the base player script

+++ Assets/Scripts/Arrow.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Arrow : MonoBehaviour
 {
-    [SerializeField] float damage = 15f;
+    [SerializeField] int damage = 15;
     Rigidbody rb;
 
     private void Awake()
